Moving SidebarNavItem component to its own file
