+++ src/renderer/components/layout/sidebar-context.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const SidebarContext = React.createContext<SidebarContextValue>({ pinned: false });
+
+export type SidebarContextValue = {
+  pinned: boolean;
+};
